pyproject.toml
[project]
name = "common_core_sorter"
version = "0.1.0"
description = "common-core-sorter using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.134.0,<1.0.0",
    "google-api-python-client>=2.131.0",
    "google-auth>=2.29.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "python-dotenv>=1.0.1"
]

[project.scripts]
common_core_sorter = "common_core_sorter.main:run"
run_crew = "common_core_sorter.main:run"
train = "common_core_sorter.main:train"
replay = "common_core_sorter.main:replay"
test = "common_core_sorter.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
